feat: add navback
remove extra from strings.xml
