+++ app/src/main/java/com/example/budgetapp/domain/usecase/CurrencyDecimals.kt
@@ -1,0 +1,4 @@
+package com.example.budgetapp.domain.usecase
+
+class CurrencyDecimals {
+}
