+++ app/src/main/java/com/example/budgetapp/presentation/screens/accounts/AccountCreation.kt
@@ -3,8 +3,8 @@ package com.example.budgetapp.presentation.screens.accounts
 import androidx.compose.runtime.Composable
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.budgetapp.presentation.screens.BaseScreen
-import com.example.budgetapp.presentation.viewmodel.AccountFormViewModel
-import com.example.budgetapp.presentation.viewmodel.AccountsViewModel
+import com.example.budgetapp.presentation.viewmodel.validationVM.AccountFormViewModel
+import com.example.budgetapp.presentation.viewmodel.dataVM.AccountsViewModel
 import com.example.budgetapp.presentation.viewmodel.uimodel.AccountUI
 
 /**
@@ -12,7 +12,10 @@ import com.example.budgetapp.presentation.viewmodel.uimodel.AccountUI
  * @param topBar - Top bar of the screen to show
  */
 @Composable
-fun AccountCreationScreen(topBar: @Composable () -> Unit) {
+fun AccountCreationScreen(
+    topBar: @Composable () -> Unit,
+    navBack: () -> Unit,
+) {
     val accountsVM = hiltViewModel<AccountsViewModel>()
     val accountFormVM = hiltViewModel<AccountFormViewModel>()
     val insertAccount = { account: AccountUI -> accountsVM.insert(account) }
@@ -21,7 +24,8 @@ fun AccountCreationScreen(topBar: @Composable () -> Unit) {
         content = { modifier ->
             AccountForm(
                 modifier = modifier,
-                onSubmit = insertAccount,
+                onSaveAccount = insertAccount,
+                navBack = navBack,
                 accountFormVM = accountFormVM,
             )
         },
