+++ app/src/main/java/com/example/budgetapp/presentation/viewmodel/dataVM/DataViewModel.kt
@@ -1,17 +1,11 @@
-package com.example.budgetapp.presentation.viewmodel
+package com.example.budgetapp.presentation.viewmodel.dataVM
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.budgetapp.domain.model.Account
-import com.example.budgetapp.domain.repository.AccountsRepository
 import com.example.budgetapp.domain.repository.Repository
-import com.example.budgetapp.presentation.viewmodel.uimodel.AccountUI
-import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
-import java.util.Currency
-import javax.inject.Inject
 
 /**
  * ViewModel for data operations. I provides CRUD operations through a repository, mapping between
@@ -80,31 +74,5 @@ abstract class DataViewModel<D, P>(private val repository: Repository<D>) : View
         viewModelScope.launch {
             repository.delete(presentationToDomain(entry))
         }
-    }
-}
-
-/**
- * ViewModel for CRUD operation on accounts, mapping between [AccountUI] and [Account]
- * @param repository - Repository to use for data operations
- */
-@HiltViewModel
-class AccountsViewModel @Inject constructor(
-    repository: AccountsRepository
-) : DataViewModel<Account, AccountUI>(repository) {
-    override fun presentationToDomain(presentation: AccountUI): Account {
-        return Account(
-            name = presentation.name,
-            balance = presentation.balance.toBigDecimal(),
-            currency = Currency.getInstance(presentation.currency)
-        )
-    }
-
-    override fun domainToPresentation(domain: Account): AccountUI {
-        return AccountUI(
-            id = domain.id,
-            name = domain.name,
-            balance = domain.balance.toString(),
-            currency = domain.currency.currencyCode
-        )
     }
 }
